Extract more common code and docs fragment for Docker Compose.
